+++ strings/snake_case.go
@@ -1,0 +1,5 @@
+package strings
+
+func SnakeCase(input string) string {
+	return ScreamingDelimited(input, '_', "", false)
+}